Add final step of onboarding process, update tests, refactor old ones, and update E2E data
- Implemented the final step of the onboarding process in `QuickFixFinalStep.kt`.
- Added comprehensive tests for the final step in `QuickFixFinalStepTest.kt`
- Refactored existing tests to improve readability and maintainability.
- Updated end-to-end test data for the e2e tests to reflect the new onboarding process.
- Deleted the `Categories` model and replaced it with `CategoryViewModel` when needed it.
- Ensured all tests pass successfully.

+++ app/src/main/java/com/arygm/quickfix/model/profile/dataFields/Review.kt
@@ -1,6 +1,6 @@
 package com.arygm.quickfix.model.profile.dataFields
 
-data class Review(val username: String, val review: String, val rating: Double) {
+data class Review(val username: String, val review: String, val rating: Float) {
   fun toFirestoreMap(): Map<String, Any> {
     return mapOf("username" to this.username, "review" to this.review, "rating" to this.rating)
   }
